refactor: remove Error context
