+++ src/App/index.tsx
@@ -1,3 +1,4 @@
+import * as Sentry from '@sentry/react';
 import { AnalyticProvider } from 'packages/analytics';
 import React from 'react';
 import { QueryClientProvider } from 'react-query';
@@ -9,14 +10,13 @@ import { Web3Wrapper } from 'clients/web3';
 import { Layout } from 'containers/Layout';
 import { AuthProvider } from 'context/AuthContext';
 import { DisableLunaUstWarningProvider } from 'context/DisableLunaUstWarning';
-import { ErrorLoggerProvider } from 'context/ErrorLogger';
 import { SuccessfulTransactionModalProvider } from 'context/SuccessfulTransactionModalContext';
 import { MuiThemeProvider } from 'theme/MuiThemeProvider';
 
 import Router from './Router';
 
 const App = () => (
-  <ErrorLoggerProvider>
+  <Sentry.ErrorBoundary>
     <Web3Wrapper>
       <QueryClientProvider client={queryClient}>
         <MuiThemeProvider>
@@ -38,7 +38,7 @@ const App = () => (
         </MuiThemeProvider>
       </QueryClientProvider>
     </Web3Wrapper>
-  </ErrorLoggerProvider>
+  </Sentry.ErrorBoundary>
 );
 
 export default App;
