+++ src/clients/api/queries/getPendingRewards/index.ts
@@ -1,6 +1,6 @@
 import BigNumber from 'bignumber.js';
+import { logError } from 'errors';
 
-import { logError } from 'context/ErrorLogger';
 import { VError } from 'errors/VError';
 import convertPriceMantissaToDollars from 'utilities/convertPriceMantissaToDollars';
 import extractSettledPromiseValue from 'utilities/extractSettledPromiseValue';
