+++ src/errors/index.ts
@@ -1,3 +1,4 @@
 export * from './VError';
 export * from './transactionErrors';
 export * from './formatVErrorToReadableString';
+export * from './logError';
