+++ src/initializeLibraries/index.ts
@@ -1,9 +1,11 @@
 import initializeBigNumber from './bigNumber';
+import initializeSentry from './sentry';
 import initializeYup from './yup';
 
 const initializeLibraries = () => {
   initializeBigNumber();
   initializeYup();
+  initializeSentry();
 };
 
 export default initializeLibraries;
